requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.4
Gunicorn==23.0.0
psycopg2-binary==2.9.9
alembic==1.13.3
pytest==8.3.3
openai==1.51.2
pytest-mock